Added a way to connect to an existing express implementation.
